[CAS-768] Moving code from QueryChannelsControllerImpl to QueryChannelsControllerStateFlow
